+++ beacon_node/network/src/service.rs
@@ -794,7 +794,7 @@ impl<T: BeaconChainTypes> NetworkService<T> {
 
     fn subscribe_to_peer_das_topics(&mut self, subscribed_topics: &mut Vec<GossipTopic>) {
         if self.subscribe_all_data_column_subnets {
-            for column_subnet in 0..self.fork_context.spec.data_column_sidecar_subnet_count {
+            for column_subnet in 0..self.fork_context.spec.data_column_sidecar_subnet_count as u64 {
                 for fork_digest in self.required_gossip_fork_digests() {
                     let gossip_kind =
                         Subnet::DataColumn(DataColumnSubnetId::new(column_subnet)).into();
@@ -814,7 +814,7 @@ impl<T: BeaconChainTypes> NetworkService<T> {
                     .local_enr()
                     .custody_subnet_count::<<T as BeaconChainTypes>::EthSpec>(
                         &self.fork_context.spec,
-                    ),
+                    ) as u64,
                 &self.fork_context.spec,
             ) {
                 for fork_digest in self.required_gossip_fork_digests() {
